+++ vscode/e2e/default_settings/dead-link.test.ts
@@ -15,7 +15,17 @@ describe('dead links', () => {
     noteSearcher = new NoteSearcherDriver(vscode);
   });
 
-  it('shows dead links on save', async () => {
+  it('shows existing dead markdown links', async () => {
+    expect(await noteSearcher.isShowingInDeadLinks('readme.md')).to.be.true;
+    expect(await noteSearcher.isShowingInDeadLinks('to/nowhere')).to.be.true;
+  });
+
+  it('shows existing dead wiki links', async () => {
+    expect(await noteSearcher.isShowingInDeadLinks('readme.md')).to.be.true;
+    expect(await noteSearcher.isShowingInDeadLinks('non_existent_note')).to.be.true;
+  });
+
+  it('shows new dead link on save', async () => {
     await vscode.openDemoDirFile('cheese.md');
     const editor = await vscode.currentEditor();
     if (!editor) { expect.fail('expected an editor to be open'); };