+++ vscode/src/dead_links/DeadLinkFinder.test.ts
@@ -21,7 +21,7 @@ describe('dead link finder, mocked filesystem', () => {
     }
   };
 
-  describe('posix paths', () => {
+  describe('posix paths, markdown links', () => {
     if (process.platform === 'win32') { return; }
 
     beforeEach(() => {
@@ -109,7 +109,62 @@ describe('dead link finder, mocked filesystem', () => {
     });
   });
 
-  describe('windows paths', () => {
+  describe('posix paths, wiki links', () => {
+    if (process.platform === 'win32') { return; }
+
+    beforeEach(() => {
+      fileSystem = tmoq.Mock.ofType<FileSystem>();
+      linkIndex = new MapLinkIndex();
+      finder = new DeadLinkFinder(linkIndex, fileSystem.object);
+    });
+
+    it('finds dead link', () => {
+      setupLinks([
+        new MockFile('/a.md', '[[blah]]')
+      ]);
+
+      const deadLinks = finder.findAllDeadLinks();
+
+      expect(deadLinks).toHaveLength(1);
+      expect(deadLinks[0].sourcePath).toBe('/a.md');
+      expect(deadLinks[0].targetPath).toBe('blah');
+    });
+
+    it('finds no dead links', () => {
+      setupLinks([
+        new MockFile('/a.md', '[[b]]'),
+        new MockFile('/b.txt', '')
+      ]);
+
+      const deadLinks = finder.findAllDeadLinks();
+
+      expect(deadLinks).toHaveLength(0);
+    });
+
+    it('links to subdirs work', () => {
+      setupLinks([
+        new MockFile('/a.md', '[[e]]'),
+        new MockFile('/b/c/e.txt', '')
+      ]);
+
+      const deadLinks = finder.findAllDeadLinks();
+
+      expect(deadLinks).toHaveLength(0);
+    });
+
+    it('links to parent dirs work', () => {
+      setupLinks([
+        new MockFile('/b/c/e.txt', '[[a]]'),
+        new MockFile('/a.md', '')
+      ]);
+
+      const deadLinks = finder.findAllDeadLinks();
+
+      expect(deadLinks).toHaveLength(0);
+    });
+  });
+
+  describe('windows paths, markdown links', () => {
     if (process.platform !== 'win32') { return; }
 
     beforeEach(() => {
@@ -196,6 +251,61 @@ describe('dead link finder, mocked filesystem', () => {
       expect(deadLinks).toHaveLength(0);
     });
   });
+
+  describe('windows paths, wiki links', () => {
+    if (process.platform !== 'win32') { return; }
+
+    beforeEach(() => {
+      fileSystem = tmoq.Mock.ofType<FileSystem>();
+      linkIndex = new MapLinkIndex();
+      finder = new DeadLinkFinder(linkIndex, fileSystem.object);
+    });
+
+    it('finds dead link', () => {
+      setupLinks([
+        new MockFile('c:\\a.md', '[[blah]]')
+      ]);
+
+      const deadLinks = finder.findAllDeadLinks();
+
+      expect(deadLinks).toHaveLength(1);
+      expect(deadLinks[0].sourcePath).toBe('c:\\a.md');
+      expect(deadLinks[0].targetPath).toBe('blah');
+    });
+
+    it('finds no dead links', () => {
+      setupLinks([
+        new MockFile('c:\\a.md', '[[b]]'),
+        new MockFile('c:\\b.txt', '')
+      ]);
+
+      const deadLinks = finder.findAllDeadLinks();
+
+      expect(deadLinks).toHaveLength(0);
+    });
+
+    it('links to subdirs work', () => {
+      setupLinks([
+        new MockFile('c:\\a.md', '[[e]]'),
+        new MockFile('c:\\b\\c\\e.txt', '')
+      ]);
+
+      const deadLinks = finder.findAllDeadLinks();
+
+      expect(deadLinks).toHaveLength(0);
+    });
+
+    it('links to parent dirs work', () => {
+      setupLinks([
+        new MockFile('c:\\b\\c\\e.txt', '[[a]]'),
+        new MockFile('c:\\a.md', '')
+      ]);
+
+      const deadLinks = finder.findAllDeadLinks();
+
+      expect(deadLinks).toHaveLength(0);
+    });
+  });
 });
 
 describe('dead link finder, real filesystem', () => {
@@ -208,13 +318,33 @@ describe('dead link finder, real filesystem', () => {
     finder = new DeadLinkFinder(linkIndex, fs);
   });
 
-  it('finds dead link in demo dir readme', async () => {
+  it('finds dead markdown link in demo dir readme', async () => {
     await linkIndex.index(_path.resolve(__dirname, '../../demo_dir'));
 
+    // act
     const deadLinks = finder.findAllDeadLinks();
 
-    expect(deadLinks).toHaveLength(1);
-    expect(deadLinks[0].sourcePath).toContain('readme.md');
-    expect(deadLinks[0].targetPath).toContain('to/nowhere');
+    // assert
+    expect(deadLinks).toHaveLength(2);
+
+    const deadMarkdownLink = deadLinks
+      .filter(d => d.targetPath.includes('to/nowhere'))[0];
+
+    expect(deadMarkdownLink.sourcePath).toContain('readme.md');
+  });
+
+  it('finds dead wiki link in demo dir readme', async () => {
+    await linkIndex.index(_path.resolve(__dirname, '../../demo_dir'));
+
+    // act
+    const deadLinks = finder.findAllDeadLinks();
+
+    // assert
+    expect(deadLinks).toHaveLength(2);
+
+    const deadMarkdownLink = deadLinks
+      .filter(d => d.targetPath.includes('non_existent_note'))[0];
+
+    expect(deadMarkdownLink.sourcePath).toContain('readme.md');
   });
 });