+++ vscode/src/dead_links/DeadLinkFinder.ts
@@ -18,14 +18,35 @@ export class DeadLinkFinder {
     const deadLinks = [];
 
     for (const file of this.linkIndex.notes()) {
-      for (const link of this.linkIndex.linksFrom(file)) {
-        const absLinkPath = toAbsolutePath(file, link);
-        if (!this.fileSystem.fileExists(absLinkPath)) {
-          deadLinks.push(new Link(file, link));
-        }
+      for (const link of this.findDeadMarkdownLinks(file)) {
+        deadLinks.push(link);
+      }
+      for (const link of this.findDeadWikiLinks(file)) {
+        deadLinks.push(link);
       }
     }
 
+    return deadLinks;
+  };
+
+  private findDeadMarkdownLinks = (fromPath: string) => {
+    const deadLinks = [];
+    for (const link of this.linkIndex.markdownLinksFrom(fromPath)) {
+      const absLinkPath = toAbsolutePath(fromPath, link);
+      if (!this.fileSystem.fileExists(absLinkPath)) {
+        deadLinks.push(new Link(fromPath, link));
+      }
+    }
+    return deadLinks;
+  };
+
+  private findDeadWikiLinks = (fromPath: string) => {
+    const deadLinks = [];
+    for (const link of this.linkIndex.wikiLinksFrom(fromPath)) {
+      if (!this.linkIndex.containsNote(link)) {
+        deadLinks.push(new Link(fromPath, link));
+      }
+    }
     return deadLinks;
   };
 }